--- samples/gcp-storage/.eslintrc.json
@@ -1,35 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es2021": true,
-        "jest":true
-    },
-    "extends": [
-        "airbnb-base"
-    ],
-    "parserOptions": {
-        "ecmaVersion": 12
-    },
-    "rules": {
-        "linebreak-style": ["error", "windows"],
-        "no-var": "error",
-        "curly": "error",
-        "eqeqeq": "error",
-        "no-multi-spaces": "error",
-        "no-use-before-define": "error",
-        "prefer-const": "error",
-        "no-multiple-empty-lines": "error",
-        "semi": "error",
-        "space-before-blocks": "error",
-        "indent": [
-            "error",
-            2,
-            {
-                "SwitchCase": 1,
-                "MemberExpression": 1,
-                "flatTernaryExpressions": false
-            }
-        ]
-    }
-}
