--- samples/gcp-storage/.prettierrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  ...require('prettier-airbnb-config'),
-  bracketSpacing: true,
-  arrowParens: 'always',
-};
