--- samples/gcp-storage/config.aws.json
@@ -1,3 +1,0 @@
-{
-    "region": "us-east-2"
-}
