--- samples/gcp-storage/config.azure.json
@@ -1,3 +1,0 @@
-{
-    "region": "East US"
-}
