+++ samples/gcp-storage/package.json
@@ -1,15 +1,11 @@
 {
-  "name": "hello-world",
+  "name": "google-storage",
   "version": "1.0.0",
   "description": "",
-  "main": "./src/hello.js",
+  "main": "./src/storageHandler.js",
   "scripts": {
     "test": "jest",
     "lint": "eslint . --ext .js",
-    "deploy:aws": "serverless deploy --config serverless-aws.yml",
-    "remove:aws": "serverless remove --config serverless-aws.yml",
-    "deploy:azure": "serverless deploy --config serverless-azure.yml",
-    "remove:azure": "serverless remove --config serverless-azure.yml",
     "deploy:google": "serverless deploy --config serverless-google.yml",
     "remove:google": "serverless remove --config serverless-google.yml",
     "pre-push": "npm run lint && npm run test"
@@ -17,8 +13,8 @@
   "author": "MultiCloud Team",
   "license": "ISC",
   "dependencies": {
-    "@multicloud/sls-aws": "^0.1.1",
-    "@multicloud/sls-azure": "^0.1.1",
+
+    "@multicloud/sls-core": "^0.1.1",
     "core": "file:./src/multicloud-gcp/multicloud/core",
     "gcp": "file:./src/multicloud-gcp/multicloud/gcp"
   },
@@ -28,9 +24,6 @@
     "eslint-plugin-import": "^2.22.1",
     "husky": "^4.3.0",
     "jest": "^26.6.3",
-    "prettier": "^2.1.2",
-    "prettier-airbnb-config": "^1.0.0",
-    "serverless-azure-functions": "^2.1.0",
     "serverless-google-cloudfunctions": "*",
     "serverless-offline": "^6.8.0"
   },
