--- samples/gcp-storage/serverless-azure.yml
@@ -1,22 +1,0 @@
-service: multicloud-hello-world-workshop
-provider:
-  name: azure
-  resourceGroup: multicloud
-  region: ${self:custom.config.region}
-  runtime: nodejs12
-custom: 
-  config: ${file(config.azure.json)}
-plugins:
-  - serverless-azure-functions
-package:
-  exclude:
-    - local.settings.json
-    - .vscode/**
-functions:
-  hello:
-    handler: src/hello.handler
-    events:
-      - http: true
-        methods:
-          - GET
-        authLevel: anonymous
