+++ samples/gcp-storage/serverless-google.yml
@@ -7,10 +7,10 @@ provider:
   project: ${self:custom.config.project_id}
   credentials: ./.gcloud/keyfile.json
   environment:
-    projectId: ${self:custom.config.project_id} 
+    projectId: ${self:custom.config.project_id}
     clientEmail: ${self:custom.config.client_email}
     privateKey: ${self:custom.config.private_key}
-custom: 
+custom:
   config: ${file(./.gcloud/keyfile.json)}
 
 frameworkVersion: '2'
@@ -31,9 +31,9 @@ plugins:
   - serverless-google-cloudfunctions
   - serverless-offline
 
-# resources:
-#   resources:
-#     - type: storage.v1.bucket
-#       name: storage-test-from-yml-12
-#       properties:
-#         location: us-central1
+resources:
+  resources:
+    - type: storage.v1.bucket
+      name: storage_sample123
+      properties:
+        location: us-central1
